+++ src/configUI.data.ts
@@ -27,6 +27,7 @@ export const enum ErmFirstCommaCommand {
     to2 = 2,
 }
 
+export type TDefProviderMethod = 'loose_mode' | 'precision_mode' | 'precision_or_loose_mode';
 type TempConfigs = {
     CodeLens: {
         showClassReference: boolean,
@@ -34,6 +35,7 @@ type TempConfigs = {
         showDevTool: boolean,
         showFileReport: boolean,
     },
+    DefinitionProviderMethod: TDefProviderMethod,
     Diag: {
         AMasterSwitch: EDiagMasterSwitch,