+++ src/tools/DeepAnalysis/getUserDefTopClassSymbol.ts
@@ -11,3 +11,15 @@ export function getUserDefTopClassSymbol(wordUP: string): CAhkClass | null {
     }
     return null;
 }
+
+export type TFullClassMap = ReadonlyMap<string, CAhkClass>;
+
+export function getAllClass(): TFullClassMap {
+    const allMap: [string, CAhkClass][] = [];
+
+    for (const { AST } of pm.getDocMapValue()) {
+        allMap.push(...getFileAllClassMap(AST));
+    }
+
+    return new Map<string, CAhkClass>(allMap);
+}